Update constraint file name
