+++ calliope/constraints/__init__.py
@@ -1,5 +1,4 @@
 from . import base
-from . import base_planning
 from . import objective
 from . import planning
 from . import optional
